requirements.txt
Django==3.1.1
django-cors-headers==3.5.0
django-environ==0.4.5
django-extensions==3.0.8
django-filter==2.3.0
django-model-utils==4.0.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
factory-boy==3.0.1
Faker==4.1.3
ipython==7.16.1
Markdown==3.0.1
pipenv==2018.11.26
virtualenv==20.0.31
Werkzeug==1.0.1
